+++ tn-pallet/scripts/tn-pallet.js
@@ -2,26 +2,34 @@
     'use strict';
 
     function initColourPicker($context) {
-        $context.find('.tnp-colour-picker').wpColorPicker({
-            defaultColor: '#000000'
+        $context.find('.tnp-colour-picker').each(function () {
+            var $input = $(this);
+
+            $input.wpColorPicker({
+                defaultColor: '#000000'
+            });
+
+            $input.closest('.wp-picker-container').addClass('wp-picker-active');
+            $input.closest('.wp-picker-container').find('.wp-picker-holder').show();
+            $input.closest('.wp-picker-container').find('.wp-color-result').attr('tabindex', '-1').hide();
         });
     }
 
     $(function () {
         var $rows = $('#tnp-palette-rows');
-        var rowTemplate = $('#tmpl-tnp-palette-row').html();
+        var cardTemplate = $('#tmpl-tnp-palette-card').html();
 
         initColourPicker($rows);
 
         $('#tnp-add-colour').on('click', function () {
-            var $row = $(rowTemplate);
+            var $card = $(cardTemplate);
 
-            $rows.append($row);
-            initColourPicker($row);
+            $rows.append($card);
+            initColourPicker($card);
         });
 
         $rows.on('click', '.tnp-remove-colour', function () {
-            $(this).closest('.tnp-palette-row').remove();
+            $(this).closest('.tnp-palette-card').remove();
         });
     });
 })(jQuery);